WIP/RFC: drm/msm/a5xx: add support for per-instance pagetables
Per-instance pagetable should be used to prevent one process
to access gpu buffers of another and increase available address space
because it won't be shared anymore.

Problem is that IOMMUs that come with a5xx don't have TTBR1 support,
that could be used to MAP kernel object at all time like on a6xx.
This add few edge cases which need to be handled with care.
One being when some app exits and pagetable needs to be destroyed we
need switch to some other pagetable (the global one). This isn't done
correctly right now.

Also this patch does some magic with copying PGD entry from kernel global
pgtable to private so we wouldn't need to map every kernel object manually.
This assumes that at the point of PP pgtable creating (when app launches
for first time) we mapped at least one kernel object (actually should be
all). Since PGD for 4k pages is 1GiB it won't be a block entry and we
map way less than that PTE won't be a block but a table. Copying the PTE
will mirror that kernel mapping (which is aligned to that 1GiB)
into private pagetables. Before PP pagetable destruction we need to
reverse this change to prevent anything bad from happening.
This approach would also work on 32-bit GPUs but would limit private address
space to 3GiB, unless we copy PMD entries instead and have all kernel object
mapped by that time.

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>