media: qcom: venus: add msm8953 hacks
on msm8953 venus is not happy with out_height not being aligned to 32
and emits change event.
The problem is that current code treat width/height from event data as
input width/height, which might (not) be true for older firmware.
But if it was input width/height we wouldn't need to change because
we already align input height. So it could be output or both
input/output dimensions.

Also we can (and should) skip reconfiguration and continue session if we
have sufficient resources.