HACK: media: qcom: venus: use 4xx core get/put/power ops for 3xx
The power domains/clocks shouldn't really be tied to HFI firmware version.
MSM8953 has 3xx HFI but has per-core clocks/power-domain (although
there is single core) which needs same handling as MSM8996 with 4xx HFI.
Of course i'm not going to change that =P and just hack that so it
works.