dm: support allocating error strings to enhance errors returned to userspace
Previously, ti->error and ti_error strings were pointing to statically
allocated memory (the .rodata section) and it was not possible to add
parameters to the error strings.

This commit makes possible to allocate error strings dynamically using
the "kasprintf" function, so we can add arbitrary parameters to the
strings.

We need to free the error string only if is allocated with kasprintf.
So, we introduce a function "dm_free_error" that tests if the string
points to the module area (and doesn't free the string if it does),
then it calls kfree_const. kfree_const detects if the string points to
the kernel .rodata section and frees the string if it is not in
.rodata.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@kernel.org>