media: i2c: ov5670: Support single-lane operation
Currently the driver always sets the driver up for dual-lane MIPI
output, but it also supports single-lane output. Add support for that by
checking the fwnode how many lanes are used and configure the necessary
registers based on that.

To achieve this we move setting register 0x3018 out of the general reg
sequence so we set it to the correct value. The pixel_rate value also
needs to be adjusted.

Signed-off-by: Luca Weiss <luca@z3ntu.xyz>

**NOTE**: No clue if this works on ACPI