FROMLIST v5 dt-bindings: regulator: qcom,usb-vbus-regulator: Mark reg as required
The regulator code needs to know the location of the register to write to
to switch on/off. Right now we have a driver that does this, a yaml that
partially describes it and no dts that uses it.

Switching on the VBUS for sm8250 shows that we haven't documented reg as a
required property, do so now.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>