ARM: dts: qcom: msm8974*: Add rpmpd power domains
Add support for the power domains found on MSM8974 which is replacing
the direct usage of pm8841_s2 and pm8841_s4 regulators. We need to do
this in one go to avoid touching the regulators from both regulator and
genpd driver.

On samsung-klte (which uses pma8084 instead of pm8841+pm8941) we need to
overwrite those resources and handle it a bit differently and use the
correct rpmpd compatible there.

Signed-off-by: Luca Weiss <luca@z3ntu.xyz>