WIP media: i2c: ov13b10: Support DOVDD and DVDD regulators
In addition to AVDD we want to also enable DOVDD and DVDD supplies, so
that the camera module can power up.

Use the regulator_bulk API like in many other sensor drivers. Since the
regulator core will use dummy regulators if none is available, we can
drop the _optional handling as well.

Signed-off-by: Luca Weiss <luca.weiss@fairphone.com>