aw88261: fix erroneous mask logic for SYSCTRL register initialization
As per the downstream code (aw882xx), we want the three bits AMPPD, PWDN,
and HMUTE to be kept to their current value in the register. The masking
logic to do that used an "OR" of the negated masks, which is wrong; it should
be and "AND". With this patch we restore the broken invariant that the final
value is assembled from complementary masks.

Signed-off-by: Alexandre Ferrieux <alexandre.ferrieux@orange.com>